+++ .formatter.exs
@@ -1,4 +1,4 @@
 # Used by "mix format"
 [
-  inputs: ["mix.exs", "{config,lib,test}/**/*.{ex,exs}"]
+  inputs: ["mix.exs", "mix_helpers.exs", "{config,lib,test}/**/*.{ex,exs}"]
 ]