+++ lib/appsignal/transaction.ex
@@ -79,15 +79,17 @@ defmodule Appsignal.Transaction do
   """
   @spec create(String.t(), atom) :: Transaction.t()
   def create(transaction_id, namespace) when is_binary(transaction_id) and is_atom(namespace) do
-    {:ok, resource} = Nif.start_transaction(transaction_id, Atom.to_string(namespace))
-    %Transaction{resource: resource, id: transaction_id}
+    if Appsignal.Config.active?() do
+      {:ok, resource} = Nif.start_transaction(transaction_id, Atom.to_string(namespace))
+      %Transaction{resource: resource, id: transaction_id}
+    end
   end
 
   if Mix.env() in [:test, :test_phoenix] do
     @spec to_map(Appsignal.Transaction.t()) :: map()
     def to_map(transaction) do
       {:ok, json} = Nif.transaction_to_json(transaction.resource)
-      Poison.decode!(json)
+      Appsignal.Json.decode!(json)
     end
   end