+++ lib/mix/tasks/appsignal.install.ex
@@ -136,13 +136,12 @@ defmodule Mix.Tasks.Appsignal.Install do
   defp link_config_file do
     IO.write("Linking config to config/config.exs: ")
 
-    config_file = Path.join("config", "config.exs")
     active_content = "\nimport_config \"#{appsignal_config_filename()}\"\n"
 
     if appsignal_config_linked?() do
       IO.puts("Success! (Already linked?)")
     else
-      case append_to_file(config_file, active_content) do
+      case append_to_file(config_file_path(), active_content) do
         :ok ->
           IO.puts("Success!")
 
@@ -152,10 +151,6 @@ defmodule Mix.Tasks.Appsignal.Install do
       end
     end
   end
-
-  defp config_file_path, do: Path.join("config", "config.exs")
-  defp appsignal_config_filename, do: "appsignal.exs"
-  defp appsignal_config_file_path, do: Path.join("config", appsignal_config_filename())
 
   # Checks if AppSignal was already linked in the main config/config.exs file.
   defp appsignal_config_linked? do
@@ -279,7 +274,8 @@ defmodule Mix.Tasks.Appsignal.Install do
       "prod" |> config_path_for_env |> File.exists?()
   end
 
-  defp config_path_for_env(env) do
-    Path.join("config", "#{env}.exs")
-  end
+  defp appsignal_config_filename, do: "appsignal.exs"
+  defp config_file_path, do: Path.join("config", "config.exs")
+  defp appsignal_config_file_path, do: Path.join("config", appsignal_config_filename())
+  defp config_path_for_env(env), do: Path.join("config", "#{env}.exs")
 end